@ambv 2017-01-04
While I wouldn't bet on Travis for benchmarking, master was consistently taking around 20 minutes to pass all tests. With this change, this is down to 11 minutes. ✨