@ambv 2017-01-04
OK, this is now much more friendly for review. Under 100 significant lines in total, just two files.