@ambv 2017-01-05
> (But really, I wish we could kill runtests.py in favor of migrating everything to pytest...)

Agreed! I see there's discussion on #1673, let me comment there on steps forward. For this pull request here, parallelizing testpythoneval.py was the lowest hanging fruit in terms of runtime savings which cuts down our waits **today** :)